Remove panic from reduce collation
This commit removes the panic call in the rendering of a reduce collation, which would be triggered if wrong input data were presented to the collation reduction operator. Instead, the strategy now adopted makes the reduce collation behave like other aggregates currently do: Upon being presented invalid accumulations or data, the operator reports an error to the logs and produces best-effort query results by ignoring the invalid input data.

The commit also includes a regression test for the panic, which checks the logs of the compute replica for an appropriate ERROR. To make the regression test ignore soft assertions, however, it was necessary to make an adaptation to the service configuration in the testing framework.