persist: remove source-example from persistcli
It was _extremely_ useful for helping sand down the rough edges when
"new persist" was getting started, but has since become something that
just needs to be maintained when persist's public API changes. With the
schema stuff coming soon, that maintenance is a bigger lift than
previous ones, so we need to decide if it's worth it. Aljoscha confirms
that it's not particularly providing value anymore, so goodbye, old
friend.