storage-client: track since frontiers for rehydration
The rehydrating client was previously ignoring `AllowCompaction`
commands which meant that new clusters were only receiving
`CreateSource` commands for the sources that the rehydrating client was
tracking.

For the most part this worked correctly, because the first time the
`AllowCompaction` command was being sent by the controller, it would
reach the cluster which would cleanup the dataflow, and then would
respond with a `DroppedIds` response that would lead to the rehydrating
client forgetting about a source, and therefore not re-sending it to new
clusters in future rehydrations.

In the not-happy case though, an `AllowCompaction` command could be sent
by the storage controller but the cluster could crash before it could
send the `DroppedIds` response. In this situation the rehydrating client
would forever believe that the source for which the `AllowCompaction`
command was lost should be running. The reason for this is that
the controller had already send the original `AllowCompaction` command,
so from its perspective there is nothing else to do, and the new cluster
that got rehydrated only heard about a `CreateSource` command so from
its perspective the source has not been dropped.

The solution implemented in this PR is for the rehydrating client to not
throw away the `AllowCompaction` commands but instead to keep them in
its state, just like we do for every other command, and re-send them any
time we want to rehydrate a client. This guarantees that no matter how
many crashes happen a cluster will eventually process all the pending
commands and will correctly generate a `DroppedIds` response, to clean
up the state of all the clients.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>