SQL: Prevent ambiguous col refs to system objects (#17545)
This commit prevents ambiguous column references in views and
materialized views with system objects. Examples of these are
`SELECT *` and `NATURAL JOIN`. We save the create SQL command, along
with the ambiguous column references, in the catalog as part of the
view definition. If a column is ever added to one of the underlying
objects in the view, then the ambiguous column references cause the
view definition to change and potentially become invalid.

For example, imagine there's a table t1 with one integer column and a
table t2 with one integer column, and a view with the definition:
  `CREATE VIEW v AS SELECT * FROM t1 UNION ALL SELECT * FROM t2`
If we add a column to t1, then the view definition changes and becomes
invalid.

Currently, we don't allow users to add columns to user objects, but we
do want to be able to add columns to system objects in between
versions. That's why we only prevent these ambiguous references when
they involve system objects. A much better approach would be to expand
these ambiguous column references to unambiguous column references. As
it turns out, this is much more complicated than it seems. See #16650.

Here is a bit of a hand wavy proof as to why we only need to check the
immediate view definition for system objects and ambiguous column
references, and not the entire dependency tree:
- A view with no object references cannot have any ambiguous column
references to a system object, because it has no system objects.
- A view with a direct reference to a system object and a `*` or
`NATURAL JOIN` will be rejected due to ambiguous column references.
- A view with system objects but no `*` or `NATURAL JOIN`s cannot have
any ambiguous column references to a system object, because all column
references are explicitly named.
- A view with `*` or `NATURAL JOIN`s, that doesn't directly reference a
system object cannot have any ambiguous column references to a system
object, because there are no system objects in the top level view and
all sub-views are guaranteed to have no ambiguous column references to
system objects.