storage: reclock resumption frontier in async worker
This PR is fixing a TODO I had left in the source operator that had to
do an awkward loop and interact with the reclock follower in order to
transalte the `IntoTime` resumption frontier that it was given to its
own timestamp type.

The motivation for this change is to slowly disentangle the operators
that read from a source from the act of reclocking. Their final form
should be timely operators that operate on a scope of their timestamp
and have no idea that their outputs are being reclocked. Once we reach
the point where reclocking is just a transformation applied on top of
the differential collection produced by the source ingestion we will be
able to push more computation before reclocking. For example there is no
reason decoding and envelope processing needs to happen in the mz time
domain, we could eagerly compute those and reclock their result.

Unfortunately this PR contains a bit of boilerplate code that I didn't
find a great way of getting rid off. Since our `IngestionDescription`
carries a `GenericSourceConnection` we *must*, at some point in the
code, perform a match statement like in this PR and then dispatch to
generic methods, but a well architected ingestion pipeline would require
this once and in the beginning of the pipeline. We're not there yet but
I'll leave that for a future date.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>