sql: fix bug in `HirRelationExpr::apply_to`
The `Map` case for the `apply_to` method currently chunks its `scalars`
in an incorrect way, because:

1. The `requires_nonexistent_column` flag is computed based on the arity
   of the lowered input.
2. The lowered input is prefixed with the `get_outer` columns.

To fix this issue, we change this part of the code to keep track of the
arity of the columns lowered so far w.r.t. the HIR input representation.
To do that, we introduce a new tracking variable (`lowered_arity`) and
use it in the `requires_nonexistent_column` check instead of `old_arity`.