persist: fix "blob inputs should be cumulative" panic in usage.
It's possible, due to races, that a writer has more referenced batches
in state than we saw for that writer in blob. Cap it at the number of
bytes we saw in blob, otherwise we could hit the "blob inputs should be
cumulative" panic below.

Ironically, I had this right in the original impl. Then, in code review,
I forgot and pulled it back out. :facepalm:

Touches #17752