bin/mz: Do not treat inability to open browser as a fatal error (#17764)
When attempting to log in on a machine that can not display a
browser (e.g. a terminal-only linux box), inability to open a browser
is not fatal: It only means the user must manually visit the web page
to authorize the CLI.

So, instead of bailing, let's show the URL and wait.