compute: report errors through subscribe sinks
Prior to this change, subscribe sinks did not know how to handle
dataflow errors in their input and simply ignored them. This could
potentially cause correctness issues, since it is not clear what
semantics a dataflow's `ok` collection has when the `err` collection is
not empty.

This commit makes the subscribe sink handle errors by piggy-backing on
the existing subscribe error reporting support originally introduced to
error on too-large results. This way of error reporting is somewhat
unsatisfactory, since it does not allow retraction of errors. At the
moment this is fine, since adapter aborts subscribes upon seeing the
first error anyway. Better subscribe error reporting is tracked in
issue #17781.