storage: fix graceful sink restarts
Before, we were not updating the `as_of` of the `StorageSinkDesc` when
internally (internal to the timely cluster/replica) restarting a sink. This
lead to "cannot serve requested as_of" panics, which would restart the
whole timely cluster/replica.

Now, we go through the async worker to update the `as_of`, similar to
how we update the resumption frontier for sources using the same path.

Fixes #17543