persist: trim down new (schema'd) parquet serialization
This abandons the `arrow2::io::parquet::write::FileWriter` wrapper in
favor of directly using the underlying `parquet2::write::FileWriter`.
This allows us to skip encoding the (redundant for our case) arrow
schema in the parquet key-val metadata as well as skip the created_by
string.

For the row::encoding::tests::parquet_roundtrip test, this lowers the
encoded size of the parquet file from 1,567 bytes to 759. This latter is
made up of 346 bytes to encode the actual columnar data and the rest are
parquet metadata.