persist: introduce a type-erased, arc-weak state-by-shard cache
Building further toward "one state per shard per process".

This introduce a cache of a sharable copy of State indexed by shard_id.
It's type erased, so the stored states can have different K, V, T, D
params. It's based on weak Arc references, so if all active uses of the
shard State (e.g. handles, maintenance) are dropped, then the cache
doesn't keep us holding the State in memory.

This also uses PersistClientCache (which is one-per-process in prod) as
an entry point to the StateCache and plumbs it down to the necessary
place in Applier::new.