adapter: use dedicated Rust API for syncing parameters
Previously, the SystemConfigBackend would use `SHOW ALL` and `ALTER
SYSTEM` to sync system parameters from LaunchDarkly to the coordinator.

Unfortunately, this does not roundtrip correctly for variables of type
`Vec<Ident>`:

    SHOW allowed_cluster_replica_sizes;
    '"1", "2"'

    ALTER SYSTEM allowed_cluster_replica_sizes = '"1", "2"';
    error: unknown replica size '"1", "2"'

This behavior is silly, but unfortunately it is required for PostgreSQL
compatibility.

This commit adds a dedicated Rust API to the coordinator client, so that
the SystemConfigBackend can set system parameters without going through
SQL's crazy escaping rules. It also does a fair amount of refactoring to
share code between the various ways of setting system configuration
parameters; the diff is best viewed with whitespace suppressed.

An alternative would have been to introduce a new variant of `ALTER
SYSTEM` that has roundtripping semantics, e.g.

    ALTER SYSTEM SANELY allowed_cluster_replica_sizes = '"1", "2"';

but this doesn't seem worth it presently, as we have no immediate plans
to move the system parameter sync to a separate process (which is why we
originally chose to use `ALTER SYSTEM` instead of adding this dedicated
Rust API).