explain: fix reported stages in `EXPLAIN OPTIMIZER TRACE`
The `optimized_plan` passed to the `trace_entry` call needs to be
converted to the inner `MirRelationExpr` reference first, otherwise
it is not presented as part of the `EXPLAIN OPTIMIZER TRACE` output.

Some of the transformatinos done as part of the `optimize/global` stage
were also not reported until now.