postgres-util: key PG table columns using attnum
We previously tracked the relative order of columns in PG tables,
but did not maintain their "monotonic ID" expressed by
pg_attributes.attnum. This made it impossible for us to detect
schema changes where a column was dropped and a column of the same
name was immediately added. However, with this change, we are
able to detect these schema drifts on restart, albeit not within
the replication stream directly.