adapter: Store role ID in session (#17962)
This commit adds the role ID of the active role to the Session.
Previously, only the role name was stored. The name is not sufficient
to uniquely identify a role. For example, if someone drops a session's
active role and then re-creates a new role with the same name, then the
session will mistakenly think that the new session belongs to them.

Additionally, this commit adds an error, that forces a user to
disconnect if their role was concurrently dropped.

Part of #11579