Add ReduceExt trait with a reduce_pair method
This commit adds a new trait to timely-util extending the Reduce capabilities of an Arranged to also include the possibility to execute a reduction pair over the same input arrangement. Reduction pairs can be useful when two output views of a reduction are necessary, e.g., one for data and one for errors.