Report SQL-level error for net-zero records with non-zero accumulation in accumulable aggregates
This commit changes the reduction of accumulable aggregates to employ a reduction pair emulating a fallible reduce operator. One component of the reduction pair computes the accumulable aggregate output, while the other checks for errors. In this commit, only the invalid accumulation error related to net-zero records with non-zero accumulation is checked. Subsequent commits will also extend the error checking to unsigned aggregations.