compute: log export-dataflow ID mapping
This commit adds a new `dataflow_id` field to the `mz_compute_exports`
introspection source. This allows mapping catalog objects, like indexes
and MVs, to dataflows and their operators.

From a UI point of view, the only change is the additional column in
`mz_compute_exports`. Under the hood I refactored the logging code a bit
to more clearly decouple dataflow exports from dataflows themselves.
Prior to this commit, dataflows and exports where treated as equivalent,
which is not actually correct as dataflows can, in principle, have
multiple exports.