Concat errors from bucketed and accumulable aggregates
This commit revisits the logic in render_reduce_plan to ensure that the error streams potentially returned from a bucketed aggregate and the error stream returned from an accumulable aggregate are both concatenated and returned into the final collection bundle for the rendered reduction. The commit fixes merge skew from changes in PR #17918.