compute: refactor compute logging for readability
This commit refactors the code for logging `ComputeEvents` for
readability. There should be no functional changes.

The most signficant change is that the bulk of the demux operator logic
is now implemented by methods of a new `DemuxHandler` struct. This has
the benefit of eliminating the insane rightward drift that existed
before, due to the several indentation levels introduced by nested
closures and control flow constructs.

Other changes include:
 * More consistent naming, e.g., "exports"/"imports" instead of
   "dataflows"/"sources".
 * `Row` encoding now happens outside the demux operator for all logging
   streams, not just most of them.
 * Demux outputs now transport datums with named fields, rather than
   tuples.
 * More comments.