persist: consolidate as we go in snapshot_and_fetch
Sometimes a snapshot will be read which consolidates really well (e.g.
when the since has been heavily held back but we're reading near the
upper). Doing something reasonable here in `persist_source` is
complicated and tracked in #16858, but for `snapshot_and_fetch` (which
is used in the controllers) it's straightforward to consolidate as we go
to keep memory usage down.