adapter: correct alteration of source/sink size
We were previously failing to update `mz_sources` and `mz_sinks` with
the new size of the source or sink. If the source or sink was later
dropped, we'd then generate an invalid retraction in `mz_sources` or
`mz_sinks`.

See the comment within the patch for details on the bug and the fix.

Fix #18061.