Invalid accumulation checks on unsigned sums in reduce_pair
This commit refactors and rectifies the changes in PRs #17709 and #16852 so that query-level errors are produced when invalid accumulations  hit unsigned integer sums. The approach introduced in this commit eliminates any need for changes in HIR-to-MIR lowering and also allows us to produce errors upon invalid accumulations of uint8 data. Additionally, the commit ensures that the rendered behavior for unsigned sums otherwise matches that of unsigned sums by preferring wraparound to error when possible.