Use error instead of soft assertions in reduce_pair
This commit changes the uses of soft_assert_or_log to simply log error inside the accumulable reduce_pair operator that handles errors. The latter can be done as it does not make sense anymore to panic debug builds when a clean query-level error is produced for the tested conditions. The commit also includes a few minor code clean-up changes.