storage: filter out dropped sources/sinks in reconciliation
Without this, we can run into cases where we have recorded an
`AllowCompaction` message in rehydration state, but we still send out a
rendering command for some non-empty resumption frontier to the worker.
The worker, when seeing this will try to render the source, which will
likely fail because the since of the remap shard might be the empty
frontier.

Now, we skip rendering those sources, but make sure we still send back
the `DroppedIds` responses.

Without my fix, we might observe a sequence like this, when rehydrating
(this is with some additional logging statements:

```
2023-03-14T10:39:57.583673Z  INFO mz_storage::storage_state: AllowCompaction for u4: Antichain { elements: [] }
2023-03-14T10:39:57.583788Z  INFO mz_storage::storage_state: dropping u4
2023-03-14T10:39:57.586006Z  INFO mz_storage::storage_state: reporting dropped IDs: [User(4)]
2023-03-14T10:39:57.623997Z  INFO mz_storage::storage_state: worker 0/1 trying to (re-)start ingestion u4 at resumption frontier Antichain { elements: [1678790391001] }
2023-03-14T10:39:57.624219Z  INFO mz_storage::source::source_reader_pipeline: timely-0 building source pipeline id=u4 resume_upper={1678790391001}
2023-03-14T10:39:57.627703Z  INFO mz_storage::source::source_reader_pipeline: timely-0 source(u4) initial resume upper: source_upper={(0..+inf, 0), (0, 12)}
2023-03-14T10:39:57.628612Z  INFO mz_storage::source::kafka: into_reader: instantiating Kafka source reader at offsets {0: 12} source_id="u4" worker_id=0 num_workers=1
2023-03-14T10:39:57.628932Z  INFO mz_storage::source::source_reader_pipeline: Health for source u4 being written to sd9c29e73-5419-4288-814d-2fa7968bb616
2023-03-14T10:39:57.628976Z  INFO mz_storage::source::source_reader_pipeline: Health transition for source u4: None -> Starting
thread 'timely:work-0' panicked at 'failpoint invalid_remap_as_of panic', /Users/aljoscha/.cargo/registry/src/github.com-1ecc6299db9ec823/fail-0.5.1/src/lib.rs:496:25
note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace
fish: Job 1, 'env FAILPOINTS=invalid_remap_as…' terminated by signal SIGABRT (Abort)
```

Notice how we know that the source is dropped, and we report that back,
but we also try and render the source, which then fails and crashes the
cluster. This is because without the fix we see the `CreateSource`
command first, which enqueues it for rendering, and then see the
`AllowCompaction` command.