adapter: Check role attributes when executing commands (#17970)
This commit updates the adapter so that it optionally checks a role's
attributes before executing a command. This feature is behind a feature
flag that any superuser can toggle. This allows users to set
attributes for all users, and only enable RBAC checks when they're
ready. They can also disable the checks if they realise that they
weren't ready or made a mistake.

Part of #11579