Allow some load generators to be monotonic sources
It is convenient for testing to be able to trigger monotonic plans
without having to setup an external source.

Counter documents that `max_cardinality = None` produces an
append-only collection.  Auction and Datums never issue any
retractions.  Tpch can be append-only in the case where the tick
interval is not set, but I leave it false, as this seems like a
reasonable selection of monotonic sources as it is.